Add confirmation prompt for force kill script - Added a check to confirm if user wants to kill all processes in the script. - This will prevent accidental termination of important processes.

+++ shutdown/force_kill.sh
@@ -2,6 +2,11 @@
 
 # force_kill.sh - force kill known problem processes
 
+if [[ "$#" -eq 1 ]]; then
+	zenity --question --text="Kill all?" || exit
+fi
+
+
 # Wine
 kill 997 1021 >/dev/null 2>&1
 wineserver -k 15 && echo "Killed wineserver" || echo "wineserver not running"
@@ -12,5 +17,10 @@ killall -9 ollama_llama_server >/dev/null 2>&1 && echo "Killed Ollama" || echo "
 # Steam
 killall -9 steamwebhelper >/dev/null 2>&1 && echo "Killed Steam" || echo "Steam not running"
 
+# Osrs
+matches=$(ps aux | grep -E "jagex" | wc -l)
+procs=$(echo $matches-1 | bc)
 
-
+for proc in $(ps aux | grep -E "jagex" | tail -5 | awk '{print $2}'); do
+	kill -9 $proc && echo "Killed PID $proc" || echo "Error killing PID $proc"
+done
